requirements.txt
baybe[chem,simulation]==0.12.2
chemprop>=2.2.0
ipykernel>=7.0.0
jupyter>=1.1.1
matplotlib>=3.10.7
numpy==1.26
openpyxl==3.1.5
pandas>=2.3.3
rdkit>=2025.3.3
scikit-learn>=1.7.2
scipy>=1.14.1
seaborn>=0.13.2
torch-pca==1.1.0
transformers
accelerate
tiktoken
sentencepiece
scikit-optimize
fastcluster==1.2.5